Moving coord-related stuff to Board
